1_spoon_e2e/requirements.txt
appium-python-client==2.11.1
pytest
selenium
allure-pytest
